Fix some more category tests, reorder project test
Copied an integration test that was in project/test_models into
project/test_integration, where is should live. Also, apparently there
were some test failures related to category factories that I didn't
realize were happening in the last commit (weird) so I fixed those tests
up too.
